placed order cart removal

+++ helpers/user-helpers.js
@@ -762,7 +762,9 @@ module.exports={
          }
 
          db.get().collection(collection.ORDER_COLLECTION).insertOne(orderObj).then((response)=>{
-          db.get().collection(collection.CART_COLLECTION).deleteOne({user:ObjectId(order.userId)})
+          if(status=='Placed'){
+            db.get().collection(collection.CART_COLLECTION).deleteOne({user:ObjectId(order.userId)})
+          }
           resolve(response.insertedId)
          })
       })
@@ -1214,6 +1216,12 @@ module.exports={
       
       resolve({ status: true })
     })
+  },
+
+
+
+  addDeletedProductToCart:(orderId)=>{
+
   },
